Methods to both classes done
