--- src/classes/book.rb
@@ -1,7 +1,0 @@
-class Book 
-    def initialize(publisher, cover_state)
-        @publisher = publisher
-        @cover_state = cover_state
-    end
-
-end
