+++ src/classes/book/book.rb
@@ -1,0 +1,16 @@
+require_relative '../item/item'
+require 'date'
+
+class Book < Item
+  attr_accessor :publisher, :cover_state
+
+  def initialize(publisher, cover_state)
+    super(publish_date, id)
+    @publisher = publisher
+    @cover_state = cover_state
+  end
+
+  def can_be_archived?
+    (Date.today - @publish_date) > 3652 or @cover_state == 'bad'
+  end
+end
