+++ src/classes/book/label.rb
@@ -1,0 +1,18 @@
+require_relative '../item/item'
+
+class Label
+  attr_accessor :title, :color
+  attr_reader :id, :items
+
+  def initialize(title, color)
+    @id = Random.rand(1000)
+    @title = title
+    @color = color
+    @items = []
+  end
+
+  def add_item(item)
+    @items << item
+    item.label = self
+  end
+end
